Implement minterms and maxterms
I might want to make the type of the functions to get these values more
strictly typed. The reason is that I am intending to write a function
with computes the normal forms from the minterms and/or maxterms.
However, there is nothing in the list of numbers that identifies itself
as one or ther other. Therefore, currently it would be possible to mistakenly
pass minterms to the conjunctive normal form generator and vice versa.